refactor(liblili2json): move hash related functions to utils.py
